+++ src/main/kotlin/com/example/reactornetdisk/repository/FolderRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import reactor.core.publisher.Flux
 
 interface FolderRepository: ReactiveCrudRepository<Folder, Long> {
-    fun findByUserId(userId: Long): Flux<Folder>
-    fun findByUserIdAndParentId(userId: Long, parentId: Long?): Flux<Folder>
-    fun deleteFolderByIdAndUserId(id: Long, userId: Long): Flux<Folder>
+    fun findByUserId(userId: Int): Flux<Folder>
+    fun findByUserIdAndParentId(userId: Int, parentId: Long?): Flux<Folder>
+    fun deleteFolderByIdAndUserId(id: Long, userId: Int): Flux<Folder>
 }
